refactor(dir): skip EEXIST check after unlinkat() on most OSes
Some OSes specifically document that they will return ENOTEMPTY in the
case of a non-empty directory, so we don't have to check for EEXIST.

Yes, this includes essentially all currently supported OSes. However, we
might later add support for OSes with different semantics, so it's
worthwhile to keep the check there.