refactor: always strip trailing slashes in helper methods
Trailing slashes don't just require that the path be a directory; they
trigger dereferencing symbolic links in contexts where they would not
normally be dereferenced. This allows escapes.

As far as I'm aware, this behavior is impossible to replicate outside
the kernel without race conditions. It *could* be replicated with known
race conditions, but that would be very difficult (it would require
another helper that heavily duplicates from open_beneath()).

Unfortunately, there doesn't seem to be a good alternative to just
ignoring trailing slashes in all the helpers (create_dir(), read_link(),
etc.). So that's what we're doing.