refactor(open): reuse symlinks accidentally opened with O_PATH
If O_PATH was specified, and O_NOFOLLOW and O_DIRECTORY were not, then
since we add in O_NOFOLLOW we might accidentally open a symlink (since
on Linux O_PATH|O_NOFOLLOW will return a file descriptor open to the
symlink).

Previously, this was handled by translating that to an ELOOP error, and
following the generic fallback code. Now, the file descriptor open to
the symlink is used to get the target of the symlink for further
resolution.